refactor: isolate agentfs surface state
